refactor: firebaseの初期化ファイルをconfigディレクトリに移動
